Frames and the Frame bundle

ManifoldDiffusions.Frame — Type

Frame{Tx, Tν, TM}

Elements of $\mathrm{F}(ℳ)$ consist of a position x::Tx on ℳ<:EmbeddedManifold and a $\mathrm{GL}(d, ℝ)$-matrix ν::Tν that consists of column vectors that form a basis for T_xℳ. All input is assumed to be in local coordinates that coincide with F( ,ℳ).

Example: A frame on the south pole on the sphere

julia> 𝕊 = Sphere(1.0)
julia> u = Frame([0. , 0.], [1. 0. ; 0.  1.], 𝕊)
julia> u.x # returns [0. , 0.]
julia> u.ν # returns [1. 0. ; 0. 1.]

ManifoldDiffusions.TangentFrame — Type

TangentFrame{Tx, Tν}

A tangent vector $(ẋ, ν̇) ∈ T_u\mathrm{F}(ℳ)$. This object consists of the frame u::Frame that it is tangent to and the velocities ẋ and ν̇.

Example:

julia> 𝕊 = Sphere(1.0)
julia> u = Frame([0. , 0.], [1. 0. ; 0.  1.], 𝕊)
julia> V = TangentFrame(u, [1. 0.] , [-0.1 0. ; -0.5 1.])

ManifoldDiffusions.Π — Function

Π(u::Frame)

Canonical projection $Π: \mathrm{F}(ℳ) → ℳ$ that maps $(x,ν)$ to $x$.

ManifoldDiffusions.Πˣ — Function

Πˣ(X::TangentFrame)

Pushforward map of the canonocal projection $Π^*: T\mathrm{F}(ℳ) → Tℳ$ that maps $(ẋ, ν̇)$ to $ẋ$

ManifoldDiffusions.Hor — Function

Hor(i::Int64, u::Frame, ℳ::TM) where {TM<:EmbeddedManifold}

Returns the horizontal vector $H_i(u)$ in $T_u\mathrm{F}(ℳ)$ as an element of type TangentFrame.

ManifoldDiffusions.FrameBundle — Type

FrameBundle

The object FrameBundle(ℳ) represents the frame bundle over a manifold ℳ.

Missing docstring.

Missing docstring for g. Check Documenter's build log for details.

ManifoldDiffusions.Hamiltonian — Function

Hamiltonian(x::Tx, p::Tp, ℳ::TM) where {Tx, Tp <: Union{AbstractArray, Real}, TM <: EmbeddedManifold}

Returns the Hamiltonian induced by the Riemannian metric for a tangent vector p to ℳ at x

Hamiltonian(u::Frame, p::TangentFrame, Fℳ::FrameBundle{TM})

Returns the Hamiltonian that results from the cometric g.

Hamiltonian(x::Tx, p::Tp, Fℳ::FrameBundle{TM})

Different representation of the Hamiltonian as functions of two vectors of size $d+d^2$

Missing docstring.

Missing docstring for Hamiltonian. Check Documenter's build log for details.

ManifoldDiffusions.Geodesic — Function

Geodesic(x₀::Tx, v₀::Tv, tt, ℳ::TM) where {Tx, Tv <: AbstractArray, TM<:EmbeddedManifold}

Returns the values of the geodesic on ℳ starting at x₀ with initial velicity v₀ on a discretized time interval tt. All input is in local coordinates.

Geodesic(u₀::Frame, v₀::TangentFrame, tt, Fℳ::FrameBundle{TM})

Returns a geodesic on Fℳ starting at u₀ with initial velocity v₀ and evaluated at a discretized time interval tt.

ManifoldDiffusions.ExponentialMap — Function

ExponentialMap(x₀::Tx, v₀::Tv, ℳ::TM) where {Tx, Tv <: AbstractArray, TM<:EmbeddedManifold}

Returns as new element of ℳ that results from $Exp_{x_0}v_0$, where the point x₀ on ℳ and initial velocity v₀ are given in local coordinates.

ExponentialMap(u₀::Frame, v₀::TangentFrame, Fℳ::FrameBundle{TM})

The exponential map on Fℳ starting from u₀ with initial velocity v₀.

Missing docstring.

Missing docstring for StochasticDevelopment. Check Documenter's build log for details.