Geodesics

A geodesic $\gamma:I\to\mathcal{M}$ with $\gamma_0=x\in\mathcal{M}$ and $\dot{γ}_0=v\in T_x\mathcal{M}$ is modelled as a solution to the Hamiltonian system

\[\mathrm{d}x = \nabla_p H(x, p)\mathrm{d}t\]

\[\mathrm{d}p = -\nabla_x H(x,p)\mathrm{d}t\]

ManifoldDiffusions.Geodesic — Function

Geodesic(x₀::Tx, v₀::Tv, tt, ℳ::TM) where {Tx, Tv <: AbstractArray, TM<:EmbeddedManifold}

Returns the values of the geodesic on ℳ starting at x₀ with initial velicity v₀ on a discretized time interval tt. All input is in local coordinates.

Geodesic(u₀::Frame, v₀::TangentFrame, tt, Fℳ::FrameBundle{TM})

Returns a geodesic on Fℳ starting at u₀ with initial velocity v₀ and evaluated at a discretized time interval tt.

ManifoldDiffusions.ExponentialMap — Function

ExponentialMap(x₀::Tx, v₀::Tv, ℳ::TM) where {Tx, Tv <: AbstractArray, TM<:EmbeddedManifold}

Returns as new element of ℳ that results from $Exp_{x_0}v_0$, where the point x₀ on ℳ and initial velocity v₀ are given in local coordinates.

ExponentialMap(u₀::Frame, v₀::TangentFrame, Fℳ::FrameBundle{TM})

The exponential map on Fℳ starting from u₀ with initial velocity v₀.

ManifoldDiffusions.ParallelTransport — Function

ParallelTransport(γ, γ̇, V₀, tt, ℳ)

returns the parallel transport of an initial vector V₀, tangent to ℳ at $\gamma_0$, along a curve γ. It is assumed γ and γ̇ are known on a discretized time interval tt