Frames and the Frame bundle

ManifoldDiffusions.Frame — Type

Frame{Tx, Tν, TM}

Elements of $\mathrm{F}(\mathcal{M})$ consist of a position x::Tx on ℳ<:EmbeddedManifold and a $\mathrm{GL}(d, \mathbb{R})$-matrix ν::Tν that consists of column vectors that form a basis for T_xℳ. All input is assumed to be in local coordinates that coincide with F( ,ℳ).

Example: A frame on the south pole on the sphere

julia> 𝕊 = Sphere(1.0)
julia> u = Frame([0. , 0.], [1. 0. ; 0.  1.], 𝕊)
julia> u.x # returns [0. , 0.]
julia> u.ν # returns [1. 0. ; 0. 1.]

ManifoldDiffusions.TangentFrame — Type

TangentFrame{Tx, Tν}

A tangent vector $(\dot{x}, \dot{\nu}) \in T_u\mathrm{F}(\mathcal{M})$. This object consists of the frame u::Frame that it is tangent to and the velocities ẋ and ν̇.

Example:

julia> 𝕊 = Sphere(1.0)
julia> u = Frame([0. , 0.], [1. 0. ; 0.  1.], 𝕊)
julia> V = TangentFrame(u, [1. 0.] , [-0.1 0. ; -0.5 1.])

ManifoldDiffusions.Π — Function

Π(u::Frame)

Canonical projection $Π: \mathrm{F}(\mathcal{M}) \to \mathcal{M}$ that maps $(x,\nu)$ to $x$.

ManifoldDiffusions.Πˣ — Function

Πˣ(X::TangentFrame)

Pushforward map of the canonocal projection $Π^*: T\mathrm{F}(\mathcal{M}) \to T\mathcal{M}$ that maps $(\dot{x}, \dot{\nu})$ to $\dot{x}$

ManifoldDiffusions.Hor — Function

Hor(i::Int64, u::Frame, ℳ::TM) where {TM<:EmbeddedManifold}

Returns the horizontal vector $H_i(u)$ in $T_u\mathrm{F}(\mathcal{M})$ as an element of type TangentFrame.

ManifoldDiffusions.FrameBundle — Type

FrameBundle

The object FrameBundle(ℳ) represents the frame bundle over a manifold $\mathcal{M}$.

ManifoldDiffusions.Hamiltonian — Function

Hamiltonian(x::Tx, p::Tp, ℳ::TM) where {Tx, Tp <: Union{AbstractArray, Real}, TM <: EmbeddedManifold}

Returns the Hamiltonian induced by the Riemannian metric for a tangent vector p to ℳ at x

Hamiltonian(u::Frame, p::TangentFrame, Fℳ::FrameBundle{TM})

Returns the Hamiltonian that results from the cometric g.

Hamiltonian(x::Tx, p::Tp, Fℳ::FrameBundle{TM})

Different representation of the Hamiltonian as functions of two vectors of size $d+d^2$

ManifoldDiffusions.StochasticDevelopment! — Function

StochasticDevelopment!(Y, W, u₀, ℳ; drift)

Simulate the process $\{U_t\}_t$ on $\mathrm{F}(\mathcal{M})$ starting at u₀ that solves the SDE $\mathrm{d}U_t = V^*(U_t)\mathrm{d}t+H(U_t) \circ \mathrm{d}W_t$ This function writes the process in Fℳ in place of Y